👌🏷️ define Profiler type in globalObject
