--- packages/rum/src/domain/profiling/types/index.ts
@@ -1,2 +1,0 @@
-export type * from './profilerApi.types'
-export type * from './rumProfiler.types'
